backend/requirements.txt
# ========== FastAPI ==========
fastapi==0.109.0
python-multipart==0.0.22
uvicorn[standard]

# ========== Configuration ==========
pydantic
pydantic-settings
python-dotenv

# ========== LangChain ==========
langchain==1.2.7
langchain-core
langchain-community
langchain-openai
langchain-postgres
langchain-tavily
langchain-text-splitters
dashscope==1.25.9
pgvector==0.3.6

# ========== LangGraph ==========
langgraph==1.0.7
langgraph-checkpoint==3.0.1
langgraph-checkpoint-postgres==3.0.1
langsmith

# ========== Database ==========
redis==5.0.1
psycopg2-binary==2.9.9

# ========== Utilities ==========
tiktoken

# ========== Testing ==========
pytest
pytest-asyncio